Use Nebula version plugins (#1067)
This replaces `gradle-consistent-versions` with `nebula.dependency-recommender` and `nebula.dependency-lock`. The purpose of this change is to enable having separate Spark 2.x and Spark 3.x modules in the build. An empty `spark3` project is included.

The dependency recommender plugin is used to get versions from `versions.props` or the per-project dependency lock files.

The dependency lock plugin is used to lock versions. Locks are now generated in the `build/` folders using `./gradlew generateLock`, and used in the build after running `./gradlew saveLock`. The JSON lock files are added in this commit. These are large because they include transitive information and a set of dependencies per configuration.

This also fixes a problem where JMH dependencies needed to be declared in the compile configuration. Now JMH dependencies can use the jmh configuration.