Update scan planning with DeleteFiles in each task (#1288)
This adds `DeleteFileIndex` to scan delete manifests and index delete files, updates `ManifestGroup` to use the index when planning tasks, and adds delete files to `FileScanTask`.

The `DeleteFileIndex` uses a map keyed by partition spec ID and partition tuple. Values of the map are sorted list of sequence numbers and corresponding `DeleteFile` instances. When looking up a `DataFile`, the potentially matching delete files are fetched using its partition tuple, then the sequence numbers are binary searched to find the matching set of delete files with sequence numbers higher than the data file.

The index also supports global equality delete files. If an equality delete file is added to the table with an unpartitioned spec, it will be returned for all data files with a lower sequence number, regardless of partition.