Actions: Add expire snapshots action (#1264)
Previously the only way to expire snapshots was through a single machine table
operation with RemoveSnapshots. In this patch we add a new Spark Action which
does the same work, but does so in a scalable way. Instead of using the old
logic for analyzing files to remove, we use the Metadata Table representations
of the table both before and after Snapshot Expiration to determine un-needed
files.